refactor(sdr-core): #596 move AcarsOutputs to acars_output.rs
Pure relocation — the struct and impl move from controller.rs's
private namespace to a pub item in acars_output.rs, where the
upcoming async-I/O refactor (#596) will grow them with a worker
thread, shared config lock, and bounded mpsc channel. No
behaviour change in this commit.
